# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: HomeNet
  description: |
    listings of properties for sale
  license:
    name: MIT
servers:
- url: http://localhost:3000
paths:
...
  /listings/{listingId}:
    get:
      summary: gets a listing by id
      operationId: getById
      tags:
      - listings
      parameters:
      - name: listingId
        in: path
        description: the listing by id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: the Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseListings"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: updates a listing
      operationId: put
      tags:
      - listings
      parameters:
      - name: listingId
        in: path
        description: the listing id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: the listing to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Listing'
      responses:
        '200':
          description: the replaced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseListings"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: deletes a listing
      operationId: deleteListing
      tags:
        - listings
      parameters:
      - name: listingId
        in: path
        description: listing id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: the deleted user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseListings"
        '204':
          description: No change (nothing was deleted)
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Listing:
      type: object
      properties:
        typeSummary:
          type: string
        price:
          type: integer
        monthlyFee:
          type: integer
        bidding:
          type: boolean
        address:
          properties:
            streetName:
              type: string
            streetNumber:
              type: string
            city:
              type: string
            coordinate:
              properties:
                lat:
                  type: integer
                lng:
                  type: integer
    ResponseListing:
      allOf:
      - $ref: "#/components/schemas/Listing"
      - properties:
          __v:
            type: integer
            format: uint32
          _id:
            type: string
            format: uuid
    ResponseListings:
      type: array
      items:
        $ref: "#/components/schemas/ResponseListing"
    Error:
      required:
        - error
      properties:
        error:
          type: object
          properties:
            expose:
              type: boolean
            statusCode:
              type: integer
              format: uint32
            status:
              type: integer
              format: uint32
            body:
              title: body as escaped json
              type: string
            type:
              title: fault desciption
              type: string